+++ cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Made transfer encoding to use configurable
+- Made content type to use configurable via `--content-type` argument
 
 
 0.1.0